Controle Ganhos

+++ main/filtro.cpp
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <math.h>
 
-int filtro(float tam, float fc, double Ak, double Afko, double *Afk){
+int filtro(double tam, float fc, double Ak, double Afko, double *Afk){
     double pi = 3.1415926;
     double T = tam/1000; //tempo de amostragem em segundos
     double a1 = 1/(1+T*2*pi*fc), a2 = T*2*pi*fc/(1+T*2*pi*fc);
